Refactored the order creation petition to other files to make it more readable
